+++ lib/screens/profile_screen.dart
@@ -15,8 +15,9 @@ import '../repositories/dependents_repository.dart';
 import '../services/dev_api.dart';
 import '../config/app_config.dart';
 
-// novo card reutilizável
+// cards reutilizáveis
 import '../ui/components/dependents_card.dart';
+import '../ui/components/visitor_profile_view.dart';
 
 const _createAccountUrl =
     'https://assistweb.ipasemnh.com.br/requerimentos/recuperar-senha-prestador';
@@ -54,7 +55,6 @@ class _ProfileScreenState extends State<ProfileScreen>
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    // Inicializa o repo uma única vez, lendo a base da AppConfig (ou dart-define)
     if (!_depsReady) {
       final baseUrl = AppConfig.maybeOf(context)?.params.baseApiUrl
           ?? const String.fromEnvironment('API_BASE', defaultValue: 'http://192.9.200.98');
@@ -84,7 +84,7 @@ class _ProfileScreenState extends State<ProfileScreen>
       final rows = await _depsRepo.listByMatricula(idMatricula);
       if (!mounted) return;
       setState(() => _deps = rows);
-    } catch (e) {
+    } catch (_) {
       if (!mounted) return;
       setState(() => _depsError = 'Falha ao carregar dependentes');
     } finally {
@@ -144,14 +144,11 @@ class _ProfileScreenState extends State<ProfileScreen>
         padding: const EdgeInsets.all(16),
         children: [
           if (_profile == null) ...[
-            _HeaderCardVisitor(
+            // <<< visão de visitante centralizada
+            VisitorProfileView(
               onLogin: _goToLogin,
               onSignUp: _openCreateAccount,
             ),
-            const SizedBox(height: 16),
-            const _LockedDataBlocks(),
-            const SizedBox(height: 24),
-            _VisitorHint(),
           ] else ...[
             _HeaderCardUser(profile: _profile!, onLogout: _logout),
             const SizedBox(height: 16),
@@ -168,93 +165,7 @@ class _ProfileScreenState extends State<ProfileScreen>
   }
 }
 
-// ================== Widgets ==================
-
-class _HeaderCardVisitor extends StatelessWidget {
-  final VoidCallback onLogin;
-  final VoidCallback onSignUp;
-
-  const _HeaderCardVisitor({
-    required this.onLogin,
-    required this.onSignUp,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Container(
-          decoration: BoxDecoration(
-            color: kCardBg,
-            borderRadius: BorderRadius.circular(16),
-            border: Border.all(color: kCardBorder, width: 2),
-          ),
-          padding: const EdgeInsets.all(16),
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              const CircleAvatar(
-                radius: 28,
-                backgroundColor: kBrand,
-                child: Icon(Icons.person_outline, color: Colors.white, size: 28),
-              ),
-              const SizedBox(width: 12),
-              const Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text('Visitante',
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                        style: TextStyle(fontSize: 18, fontWeight: FontWeight.w800)),
-                    SizedBox(height: 4),
-                    Wrap(
-                      spacing: 8,
-                      runSpacing: 6,
-                      children: [
-                        _StatusChip(
-                          label: 'Logado como Visitante',
-                          color: Color(0xFFB54708),
-                          bg: Color(0xFFFFF4E5),
-                        ),
-                        _StatusChip(
-                          label: 'Acesso limitado',
-                          color: Color(0xFF6941C6),
-                          bg: Color(0xFFF4EBFF),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(height: 12),
-        SizedBox(
-          width: double.infinity,
-          height: 44,
-          child: FilledButton.icon(
-            style: FilledButton.styleFrom(backgroundColor: kBrand),
-            onPressed: onLogin,
-            icon: const Icon(Icons.login),
-            label: const Text('Fazer login'),
-          ),
-        ),
-        const SizedBox(height: 8),
-        SizedBox(
-          width: double.infinity,
-          height: 44,
-          child: OutlinedButton.icon(
-            onPressed: onSignUp,
-            icon: const Icon(Icons.person_add_alt_1_outlined),
-            label: const Text('Criar conta'),
-          ),
-        ),
-      ],
-    );
-  }
-}
+// ================== Widgets (apenas os da visão logada) ==================
 
 class _HeaderCardUser extends StatelessWidget {
   final Profile profile;
@@ -284,10 +195,12 @@ class _HeaderCardUser extends StatelessWidget {
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Text(profile.nome,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(fontSize: 18, fontWeight: FontWeight.w800)),
+                Text(
+                  profile.nome,
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                  style: const TextStyle(fontSize: 18, fontWeight: FontWeight.w800),
+                ),
                 const SizedBox(height: 4),
                 const Wrap(
                   spacing: 8,
@@ -340,7 +253,7 @@ class _UserDataBlocks extends StatelessWidget {
           title: 'Dados do usuário',
           child: Column(
             children: [
-              _InfoRow(label: 'Nome completo', value: profile.nome),
+              _InfoRow(label: 'Nome completo ', value: profile.nome),
               _InfoRow(label: 'CPF', value: fmtCpf(profile.cpf)),
               _InfoRow(label: 'Matrícula', value: profile.id.toString()),
               _InfoRow(label: 'E-mail', value: profile.email ?? '—'),
@@ -349,8 +262,7 @@ class _UserDataBlocks extends StatelessWidget {
           ),
         ),
         const SizedBox(height: 12),
-
-        /// >>> Usa o card reutilizável
+        // Card de dependentes (reutilizável)
         DependentsCard(
           items: dependentes ?? const [],
           isLoading: depsLoading,
@@ -358,124 +270,8 @@ class _UserDataBlocks extends StatelessWidget {
           compact: true,
           showDivider: true,
           showMatricula: true,
-          // onTap: (d) { /* abrir detalhes, se desejar */ },
         ),
       ],
-    );
-  }
-}
-
-class _LockedDataBlocks extends StatelessWidget {
-  const _LockedDataBlocks();
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: const [
-        SectionCard(
-          title: 'Dados do usuário',
-          child: Column(
-            children: [
-              _LockedInfoRow(label: 'Nome completo'),
-              _LockedInfoRow(label: 'CPF'),
-              _LockedInfoRow(label: 'Matrícula'),
-              _LockedInfoRow(label: 'E-mail'),
-              _LockedInfoRow(label: 'Telefone'),
-              _LockedInfoRow(label: 'Data de nascimento'),
-              _LockedInfoRow(label: 'Vínculo / Situação'),
-            ],
-          ),
-        ),
-        SizedBox(height: 12),
-        SectionCard(
-          title: 'Benefícios',
-          child: Column(
-            children: [
-              _LockedInfoRow(label: 'Plano de saúde'),
-              _LockedInfoRow(label: 'Dependentes'),
-              _LockedInfoRow(label: 'Autorizações recentes'),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _VisitorHint extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(12),
-      decoration: BoxDecoration(
-        color: kPanelBg,
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: kPanelBorder),
-      ),
-      child: const Text(
-        'Você está logado como Visitante. Faça login para visualizar seus dados '
-            'pessoais e informações de benefícios.',
-        style: TextStyle(color: Color(0xFF475467)),
-      ),
-    );
-  }
-}
-
-class _StatusChip extends StatelessWidget {
-  final String label;
-  final Color color;
-  final Color bg;
-
-  const _StatusChip({required this.label, required this.color, required this.bg});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 4),
-      decoration: BoxDecoration(
-        color: bg,
-        borderRadius: BorderRadius.circular(999),
-        border: Border.all(color: bg),
-      ),
-      child: Text(
-        label,
-        maxLines: 1,
-        overflow: TextOverflow.ellipsis,
-        style: TextStyle(
-          color: color,
-          fontSize: 12,
-          fontWeight: FontWeight.w700,
-        ),
-      ),
-    );
-  }
-}
-
-class _LockedInfoRow extends StatelessWidget {
-  final String label;
-
-  const _LockedInfoRow({required this.label});
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      dense: true,
-      contentPadding: EdgeInsets.zero,
-      leading: const Icon(Icons.lock_outline, color: Color(0xFF667085)),
-      title: Text(
-        label,
-        maxLines: 1,
-        overflow: TextOverflow.ellipsis,
-        style: const TextStyle(fontWeight: FontWeight.w600, color: Color(0xFF101828)),
-      ),
-      subtitle: const Text(
-        'Disponível após login',
-        maxLines: 1,
-        overflow: TextOverflow.ellipsis,
-        style: TextStyle(color: Color(0xFF667085)),
-      ),
-      trailing: const Icon(Icons.chevron_right, color: Color(0xFF98A2B3)),
-      minLeadingWidth: 0,
     );
   }
 }
@@ -505,6 +301,32 @@ class _InfoRow extends StatelessWidget {
         style: const TextStyle(color: Color(0xFF475467)),
       ),
       minLeadingWidth: 0,
+    );
+  }
+}
+
+class _StatusChip extends StatelessWidget {
+  final String label;
+  final Color color;
+  final Color bg;
+
+  const _StatusChip({required this.label, required this.color, required this.bg});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 4),
+      decoration: BoxDecoration(
+        color: bg,
+        borderRadius: BorderRadius.circular(999),
+        border: Border.all(color: bg),
+      ),
+      child: Text(
+        label,
+        maxLines: 1,
+        overflow: TextOverflow.ellipsis,
+        style: TextStyle(color: color, fontSize: 12, fontWeight: FontWeight.w700),
+      ),
     );
   }
 }
